+++ napari-hub-commons/src/nhcommons/models/github_activity.py
@@ -6,7 +6,7 @@ from dateutil.relativedelta import relativedelta
 from typing import Dict, Any, List, Optional, Iterator
 
 from pynamodb.attributes import UnicodeAttribute, NumberAttribute
-from nhcommons.models.helper import set_ddb_metadata, PynamoWrapper
+from nhcommons.models.pynamo_helper import set_ddb_metadata, PynamoWrapper
 
 logger = logging.getLogger(__name__)
